+++ src/explorers/DtonsEq.c
@@ -236,9 +236,9 @@ static inline void modify_bounds(Constraints *constraints, int i, double aij,
 #ifndef TESTING
 static inline
 #endif
-    Old_and_new_coeff
-    update_row_A_dton(Matrix *A, int i, int q, int j, int k, double aij, double aik,
-                      int *row_size, PostsolveInfo *postsolve_info)
+    Old_and_new_coeff update_row_A_dton(Matrix *A, int i, int q, int j, int k,
+                                        double aij, double aik, int *row_size,
+                                        PostsolveInfo *postsolve_info)
 {
     int ii, start, end, insertion;
     double old_val = 0.0;
@@ -309,9 +309,9 @@ static inline
     //          remove variable that is substituted
     // -------------------------------------------------------------------
     memmove(A->x + subst_idx, A->x + subst_idx + 1,
-            (end - subst_idx - 1) * sizeof(double));
+            ((size_t) (end - subst_idx - 1)) * sizeof(double));
     memmove(A->i + subst_idx, A->i + subst_idx + 1,
-            (end - subst_idx - 1) * sizeof(int));
+            ((size_t) (end - subst_idx - 1)) * sizeof(int));
     end -= 1;
     diff_row_size += 1;
     insertion -= (subst_idx < insertion);
@@ -323,9 +323,9 @@ static inline
     if (ABS(new_val) <= ZERO_TOL)
     {
         memmove(A->x + insertion, A->x + insertion + 1,
-                (end - insertion - 1) * sizeof(double));
+                ((size_t) (end - insertion - 1)) * sizeof(double));
         memmove(A->i + insertion, A->i + insertion + 1,
-                (end - insertion - 1) * sizeof(int));
+                ((size_t) (end - insertion - 1)) * sizeof(int));
         end -= 1;
         diff_row_size += 1;
     }
@@ -349,9 +349,9 @@ static inline
         else
         {
             memmove(A->x + insertion + 1, A->x + insertion,
-                    (end - insertion) * sizeof(double));
+                    ((size_t) (end - insertion)) * sizeof(double));
             memmove(A->i + insertion + 1, A->i + insertion,
-                    (end - insertion) * sizeof(int));
+                    ((size_t) (end - insertion)) * sizeof(int));
             A->x[insertion] = new_val;
             A->i[insertion] = j;
             end += 1;
@@ -364,8 +364,8 @@ static inline
     A->nnz -= diff_row_size;
 
     assert(*row_size == end - start);
-    DEBUG(ASSERT_INCREASING_I(A->i + start, *row_size););
-    DEBUG(ASSERT_NO_ZEROS_D(A->x + start, *row_size););
+    DEBUG(ASSERT_INCREASING_I(A->i + start, (size_t) *row_size););
+    DEBUG(ASSERT_NO_ZEROS_D(A->x + start, (size_t) *row_size););
     assert(A->p[q].end <= A->p[q + 1].start);
     return old_and_new_coeff;
 }
@@ -736,7 +736,7 @@ static PresolveStatus remove_dton_eq_rows__(Problem *prob, int max_shift_per_row
     {
         DEBUG(verify_no_duplicates_sort_ptr(iwork_n_rows, n_new_dton_rows));
         iVec_append_array(constraints->state->dton_rows, iwork_n_rows,
-                          n_new_dton_rows);
+                          (size_t) n_new_dton_rows);
     }
 
     return status;
